review: 修复深度审查发现的问题(2 安全 + 一批健壮性)+ 补测试与消毒器 CI
四个并行审查 agent 扫全树,修掉所有真实问题(威胁模型:.spk 可能来自任何人):

安全(HIGH):
- edit 的 POSIX shell 注入:扩展名取自 pack 内文件名进了 system() 单引号命令,
  文件名 x.txt';id;' 可执行任意命令。改为只接受 [A-Za-z0-9._-] 扩展名 + 对
  路径做单引号转义(Win32 不受影响)。
- web /api/get 的 Content-Disposition 头注入:pack basename 未转义进 filename=,
  含 CR/LF/NUL 可注入响应头(httplib CRLF 检查在 NUL 处停止)。加 safe_filename()
  清洗 + /api/get,/api/view 加 nosniff + token 常时比较。

健壮性(中):
- compact() 在 rename/reopen 失败时会把 store 置空 → 后续调用崩溃;改为失败时
  重开原文件恢复。
- get_bytes 的 p+len>n 在 32 位上溢出绕过越界检查 → 改 len>n-p;加
  _FILE_OFFSET_BITS=64(32 位 off_t 上 >2GiB pread/pwrite);C ABI entry.size
  size_t→uint64_t。

防御性(低):read_decrypt 校验 enc_size==plain_size+40;POSIX rename 做目录
fsync(与 Win32 WRITE_THROUGH 对齐);sealpack_list malloc 溢出保护;
sealpack_merge 拦 target==source;C 头文档说明 list/stat 的 const char* 失效时机。
CLI:merge 文件数排除 .spkdel、rekey 后同步 g_session_password、merge -d 无源报错。

测试/CI:新增 cli ctest(cli_smoke.cmake 跨平台驱动真实二进制 create/add/ls/cat/
get/merge+.spkdel);test_capi 覆盖 sealpack_merge;CI 加 ASan+UBSan job。本地
ASan+UBSan 全 8 项通过、Release 全 8 项通过。

+++ include/sealpack.h
@@ -101,8 +101,10 @@ int sealpack_commit(sealpack_t* sp);
 // ---- listing / maintenance --------------------------------------------------
 
 typedef struct {
-    const char* path;   // owned by sp — valid until close/next mutation
-    size_t      size;   // plaintext size
+    const char* path;   // owned by sp — invalidated by close, the next mutation, OR
+                        // the next sealpack_list/sealpack_stat on the same handle
+                        // (they reuse one cache). Copy it if you need to keep it.
+    uint64_t    size;   // plaintext size (u64: a logical file may exceed 4 GiB)
     uint64_t    mtime;  // last-written unix seconds (UI "modified" column)
 } sealpack_entry;
